Stop ignoring .meta files
When tried to use package without .meta files you got error like that:
Asset Packages/com.graesonb.chat-gpt-wrapper-for-unity/Chat.cs has no meta file, but it's in an immutable folder. The asset will be ignored.
Ignoring asset files in common is not a good idea, it breaks internal links between assets and leads to such errors